Docker/requirements.txt
numpy==1.26.4
pandas==2.2.2
keras==3.8.0
tensorboard==2.18.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow-cpu==2.18.0
tensorflow-estimator>=2.15.0
tensorflow-hub==0.16.0
tensorflow-io-gcs-filesystem==0.35.0
anvil-uplink==0.3.36
boto==2.49.0
boto3==1.34.12
pytz==2025.1
python-dateutil==2.9.0.post0
protobuf<4.24,>=3.19.6
grpcio>=1.48.2
google-auth-oauthlib<2,>=0.5
markdown>=2.6.8
requests<3,>=2.21.0
absl-py>=0.4
h5py==3.11.0